Update readme to work with newest Docker library
When docker ditched the `docker-py` library in favor of `docker` on PyPI they made a few very very small changes that break this library.

I've gone ahead and upgraded the readme to work with these new changes. I should point out that these changes are trivial- it basically amounts to changing `docker.client()` to `docker.APIClient()`. That's because the "low level API" that this project is based on has moved- but it is still there, and completely compatible with this library. Literally all anyone has to do in order to use the latest docker python library is change that one call so they get the low level client instead of the high level one.